Fix format for sol files

+++ core/tests/ts-integration/evm-contracts/ERC20.sol
@@ -2,14 +2,14 @@
 
 pragma solidity >=0.7.0;
 
-contract ERC20{
+contract ERC20 {
     string public symbol;
-    string public  name;
+    string public name;
     uint8 public decimals;
-    uint public totalSupply;
+    uint256 public totalSupply;
 
-    mapping(address => uint) balances;
-    mapping(address => mapping(address => uint)) allowed;
+    mapping(address => uint256) balances;
+    mapping(address => mapping(address => uint256)) allowed;
 
     event Transfer(address indexed _from, address indexed _to, uint256 _value);
     event Approval(address indexed _owner, address indexed _spender, uint256 _value);
@@ -23,25 +23,28 @@ contract ERC20{
         emit Transfer(address(0), msg.sender, totalSupply);
     }
 
-
-    function balanceOf(address tokenOwner) public view returns (uint balance) {
+    function balanceOf(address tokenOwner) public view returns (uint256 balance) {
         return balances[tokenOwner];
     }
 
-    function transfer(address to, uint tokens) public returns (bool success) {
+    function transfer(address to, uint256 tokens) public returns (bool success) {
         balances[msg.sender] = safeSub(balances[msg.sender], tokens);
         balances[to] = safeAdd(balances[to], tokens);
         emit Transfer(msg.sender, to, tokens);
         return true;
     }
 
-    function approve(address spender, uint tokens) public returns (bool success) {
+    function approve(address spender, uint256 tokens) public returns (bool success) {
         allowed[msg.sender][spender] = tokens;
         emit Approval(msg.sender, spender, tokens);
         return true;
     }
 
-    function transferFrom(address from, address to, uint tokens) public returns (bool success) {
+    function transferFrom(
+        address from,
+        address to,
+        uint256 tokens
+    ) public returns (bool success) {
         balances[from] = safeSub(balances[from], tokens);
         allowed[from][msg.sender] = safeSub(allowed[from][msg.sender], tokens);
         balances[to] = safeAdd(balances[to], tokens);
@@ -49,28 +52,27 @@ contract ERC20{
         return true;
     }
 
-    function allowance(address tokenOwner, address spender) public view returns (uint remaining) {
+    function allowance(address tokenOwner, address spender) public view returns (uint256 remaining) {
         return allowed[tokenOwner][spender];
     }
 
-    function safeAdd(uint a, uint b) internal pure returns (uint c) {
+    function safeAdd(uint256 a, uint256 b) internal pure returns (uint256 c) {
         c = a + b;
         require(c >= a);
     }
 
-    function safeSub(uint a, uint b) internal pure returns (uint c) {
+    function safeSub(uint256 a, uint256 b) internal pure returns (uint256 c) {
         require(b <= a);
         c = a - b;
     }
 
-    function safeMul(uint a, uint b) internal pure returns (uint c) {
+    function safeMul(uint256 a, uint256 b) internal pure returns (uint256 c) {
         c = a * b;
         require(a == 0 || c / a == b);
     }
 
-    function safeDiv(uint a, uint b) internal pure returns (uint c) {
+    function safeDiv(uint256 a, uint256 b) internal pure returns (uint256 c) {
         require(b > 0);
         c = a / b;
     }
-
 }
